requirements.txt
# requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0
PyYAML>=5.4.0
ultralytics  # yolov5
matplotlib>=3.4.0
tqdm>=4.65.0
detectron2==0.6  #version freeze
 # detectron2 @ git+https://github.com/facebookresearch/detectron2.git